requirements.txt
event_stream
matplotlib
numpy
tonic
pillow
scikit-image
